refactor(instance-detail): reduce complexity by moving operation types to a new file
